Implementation of Strategic Management Information (Part B)
Configured the data structures, access controls, business rules and workflows as per the requirements in the Salesforce. Created a web integration solution to allow the capture of data from a web page into the system. Built report and management dashboards to meet the analytical requirements. The necessary screenshots of salesforce are included in the report.